requirements.txt
Flask~=3.0
Flask-SQLAlchemy~=3.1
Flask-JWT-Extended~=4.6
Flask-Limiter~=3.8
redis~=5.0
psycopg[binary]~=3.1